+++ graphics/java/android/graphics/Rasterizer.java
@@ -27,7 +27,7 @@ public class Rasterizer {
         finalizer(native_instance);
     }
 
-    private static native void finalizer(int native_instance);
+    private static native void finalizer(long native_instance);
 
-    int native_instance;
+    long native_instance;
 }
